refactored options set check in health.
