+++ app/DynamicsAdapter/Fams3Adapter.Dynamics/Types/BaseTypeService.cs
@@ -1,0 +1,20 @@
+﻿using Fams3Adapter.Dynamics.OptionSets.Models;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Fams3Adapter.Dynamics.Types
+{
+   public static  class TypeService
+    {
+
+        public static readonly List<string> TypeList = new List<string>()
+        {
+            {"ssg_canadianprovincecodesimpletype" },
+            {"ssg_identificationtypes" },
+            {"ssg_informationsourcecodes"},
+            {"ssg_addresscategorycodes" },
+           
+        };
+    }
+}
