+++ app/DynamicsAdapter/Fams3Adapter.Dynamics/Types/LocationType.cs
@@ -3,14 +3,15 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Fams3Adapter.Dynamics.Address
+namespace Fams3Adapter.Dynamics.Types
 {
+    //ssg_addresscategorycodes
     public class LocationType : Enumeration
     {
         public static LocationType Residence = new LocationType(867670001, "Residence");
         public static LocationType Mailing = new LocationType(867670000, "Mailing");
 
-        protected LocationType(int value, string name) : base(value, name)
+        public LocationType(int value, string name) : base(value, name)
         {
         }
